requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
supabase>=2.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.2.0
pandas>=2.1.0
